List update is working. Error responses still need work

+++ src/main/java/com/isel/daw/checklist/repositories/CheckListRepository.java
@@ -4,6 +4,5 @@ import com.isel.daw.checklist.model.CheckList;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CheckListRepository extends JpaRepository<CheckList,Integer>{
-
-    CheckList findById(int id);
+    CheckList findById(long id);
 }
